+++ src/frontend/js/components/DashboardCreditCardsManagement/index.spec.tsx
@@ -20,6 +20,7 @@ import { DashboardTest } from 'components/Dashboard/DashboardTest';
 import { DashboardPaths } from 'utils/routers/dashboard';
 import { CreditCard } from 'types/Joanie';
 import { confirm } from 'utils/indirection/window';
+import { expectBreadcrumbsToEqualParts } from 'utils/test/expectBreadcrumbsToEqualParts';
 
 jest.mock('utils/context', () => ({
   __esModule: true,
@@ -73,6 +74,7 @@ describe('<DashboardCreditCardsManagement/>', () => {
     });
     // No error is shown.
     expect(screen.queryByText('An error occurred', { exact: false })).toBeNull();
+    expectBreadcrumbsToEqualParts(['Back', 'My preferences']);
     // The empty placeholder is shown.
     screen.getByText("You haven't created any credit cards yet.");
   });