+++ src/frontend/js/utils/test/expectUrlMatchLocationDisplayed.ts
@@ -1,0 +1,6 @@
+import { screen } from '@testing-library/react';
+
+export const expectUrlMatchLocationDisplayed = (url: string) => {
+  // expect(screen.getByTestId('location-display').textContent).toMatch(new RegExp(url + '$'));
+  screen.getByTestId('location-display-' + url);
+};